modules/03-pgvector/requirements.txt
# Python Dependencies for Clean Function-Based Semantic Search
# Optimized for PostgreSQL 17 + pgvector with modular architecture

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and vector operations
psycopg2-binary==2.9.7
pgvector==0.2.4

# Machine learning and embeddings (optional - fallback to TF-IDF if not available)
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# Scientific computing
numpy>=1.26.0
scikit-learn==1.3.2

# Data validation and serialization
pydantic==2.5.0

# HTTP client and utilities
httpx==0.25.2
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0

# Production utilities
python-dotenv==1.0.0
rich==13.7.0

# Optional: Analysis and visualization
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
pytest>=7.4.0             # Testing framework
pytest-asyncio>=0.21.0    # Async testing support
httpx>=0.25.0             # HTTP client for testing APIs
